aualizacao

+++ Semestre 02/Banco de Dados/21.02.20Aula03/ExercicioAula03.sql
@@ -5,7 +5,7 @@ CREATE TABLE tb_departamento(
 );
 
 CREATE TABLE tb_projeto(
-    cd_projeto NUMBER(5),
+    cd_projeto INTEGER,
     nm_projeto VARCHAR2(300),
     dt_ini_projeto DATE,
     dt_fin_projeto DATE,
